+++ world_gen/src/chunk/height_map.rs
@@ -1,11 +1,9 @@
 use std::cmp::Ordering;
 
-use utility::Float;
-
 pub struct HeightMap {
     size: i32,
     resolution: i32,
-    height_list: Vec<Float>
+    height_list: Vec<f64>
 }
 
 impl HeightMap {
@@ -22,17 +20,17 @@ impl HeightMap {
         }
     }
 
-    pub fn get_interpolated_height(&self, relative_pos: [Float; 2]) -> Float {
+    pub fn get_interpolated_height(&self, relative_pos: [f64; 2]) -> f64 {
         let root_pos = [clamp((relative_pos[0].floor() as i32) / self.resolution, 0, self.size - 1),
                         clamp((relative_pos[1].floor() as i32) / self.resolution, 0, self.size - 1)];
-        let reference_height: [Float; 4] = [
+        let reference_height: [f64; 4] = [
             self.get(&root_pos),
             self.get(&[i32::min(root_pos[0] + 1, self.size - 1), root_pos[1]]),
             self.get(&[root_pos[0], i32::min(root_pos[1] + 1, self.size - 1)]),
             self.get(&[i32::min(root_pos[0] + 1, self.size - 1), i32::min(root_pos[1] + 1, self.size - 1)])
         ];
-        let relative_point = [relative_pos[0] - root_pos[0] as Float,
-                              relative_pos[1] - root_pos[1] as Float];
+        let relative_point = [relative_pos[0] - root_pos[0] as f64,
+                              relative_pos[1] - root_pos[1] as f64];
 
         let res = interpolate(relative_point, reference_height);
         res
@@ -47,7 +45,7 @@ impl HeightMap {
         self.resolution
     }
 
-    pub fn get_min(&self) -> Float {
+    pub fn get_min(&self) -> f64 {
         match self.height_list.iter().min_by(|a, b| {
             if a > b {
                 Ordering::Greater
@@ -62,7 +60,7 @@ impl HeightMap {
         }
     }
 
-    pub fn get_max(&self) -> Float {
+    pub fn get_max(&self) -> f64 {
         match self.height_list.iter().max_by(|a, b| {
             if a > b {
                 Ordering::Greater
@@ -77,27 +75,27 @@ impl HeightMap {
         }
     }
 
-    pub fn set(&mut self, pos: &[i32; 2], height: Float) {
+    pub fn set(&mut self, pos: &[i32; 2], height: f64) {
         let index = self.calculate_index(pos);
         self.height_list[index] = height;
     }
 
-    pub fn get(&self, pos: &[i32; 2]) -> Float {
+    pub fn get(&self, pos: &[i32; 2]) -> f64 {
         self.height_list[self.calculate_index(pos)]
     }
     #[allow(unused)]
-    pub fn set_by_index(&mut self, index: usize, height: Float) {
+    pub fn set_by_index(&mut self, index: usize, height: f64) {
         self.height_list[index] = height;
     }
 
     #[allow(unused)]
-    pub fn get_by_index(&self, index: usize) -> Float {
+    pub fn get_by_index(&self, index: usize) -> f64 {
         debug_assert!(index < self.height_list.len());
         self.height_list[index]
     }
 
     #[allow(unused)]
-    fn get_quad_heights(&self, anchor: [i32; 2]) -> [Float; 4] {
+    fn get_quad_heights(&self, anchor: [i32; 2]) -> [f64; 4] {
         [self.get(&anchor),
          self.get(&[anchor[0] + 1, anchor[1]]),
          self.get(&[anchor[0], anchor[1] + 1]),
@@ -111,14 +109,14 @@ impl HeightMap {
     }
 }
 
-fn interpolate(p: [Float; 2], reference: [Float; 4]) -> Float {
+fn interpolate(p: [f64; 2], reference: [f64; 4]) -> f64 {
     let anchor = [p[0].floor() as i32, p[1].floor() as i32];
-    let a = anchor[0] as Float + 1. - p[0];
-    let b = p[0] - anchor[0] as Float;
+    let a = anchor[0] as f64 + 1. - p[0];
+    let b = p[0] - anchor[0] as f64;
     let r_1 = a * reference[0] + b * reference[1];
     let r_2 = a * reference[2] + b * reference[3];
-    let c = anchor[1] as Float + 1. - p[1];
-    let d = p[1] - anchor[1] as Float;
+    let c = anchor[1] as f64 + 1. - p[1];
+    let d = p[1] - anchor[1] as f64;
     c * r_1 + d * r_2
 }
